pyproject.toml
[tool.poetry]
name = "Library Management API"
version = "0.1.0"
description = ""
authors = ["bcx <femolak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
pydantic = "^2.9.1"
pytest = "^8.3.3"
motor = "3.5.1"
sqlalchemy = "2.0.34"
psycopg2-binary = "2.9.9"
alembic = "^1.13.2"
bcrypt = "^4.2.0"
httpx = "^0.27.2"
aio-pika = "^9.4.3"
asyncpg = "^0.29.0"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
